extract BOM combox box.
